move tests which need to stay in k8s.io/kubernetes for API scheme
