+++ pkg/apimachinery/tests/api_meta_help_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package meta_test
+package tests
 
 import (
 	"reflect"
