+++ pkg/apimachinery/tests/apis_meta_v1_unstructed_unstructure_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package unstructured_test
+package tests
 
 import (
 	"fmt"
