+++ pkg/apimachinery/tests/runtime_helper_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package runtime_test
+package tests
 
 import (
 	"testing"
