+++ pkg/apimachinery/tests/watch_until_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package watch_test
+package tests
 
 import (
 	"errors"
