+++ pkg/apis/meta/v1/unstructured/BUILD
@@ -5,7 +5,6 @@ licenses(["notice"])
 load(
     "@io_bazel_rules_go//go:def.bzl",
     "go_library",
-    "go_test",
 )
 
 go_library(
@@ -19,22 +18,6 @@ go_library(
         "//pkg/types:go_default_library",
         "//pkg/util/json:go_default_library",
         "//vendor:github.com/golang/glog",
-    ],
-)
-
-go_test(
-    name = "go_default_xtest",
-    srcs = ["unstructured_test.go"],
-    tags = ["automanaged"],
-    deps = [
-        "//pkg/api:go_default_library",
-        "//pkg/api/testapi:go_default_library",
-        "//pkg/api/validation:go_default_library",
-        "//pkg/apimachinery/registered:go_default_library",
-        "//pkg/apis/meta/v1:go_default_library",
-        "//pkg/apis/meta/v1/unstructured:go_default_library",
-        "//pkg/runtime:go_default_library",
-        "//pkg/types:go_default_library",
     ],
 )
